Use verify to test calls to a method and use ArgumentCaptor to test the arguments passed to a method.
verify can be used to check how many times a method is used, not used or stuff related to the calling of that method.
ArgumentCaptor is used to get the values of the arguments passed to the method that is under test. We can then get the values of the arguments passed to it.